Implements order payment processing
Adds functionality to mark orders as paid.

This change introduces a new 'PAID' status to the order lifecycle,
allowing orders to be explicitly marked as paid after they have been
closed. A new endpoint is added to facilitate this transition.

The order must be in CLOSED status before it can be marked as PAID.